requirements.txt
# Core Security Components
wazuh-agent==4.7.0
bitdefender-sdk==2.1.0
action1-api==1.2.0

# Network Security
suricata==6.0.10
rita==3.3.1
packetfence==12.0.0
nvis-client==2.0.0

# Monitoring & Analysis
splunk-sdk==1.7.3
thehive4py==1.8.1

# Deception Technology
honeypy==0.6.2
canarytokens==2.1.0

# Integration & Automation
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Development Tools
pytest==7.4.3
black==23.12.1
flake8==7.0.0
